Move `get_category_id_map()` into template-utils and make it a template_global
