+++ nyaa/views/torrents.py
@@ -9,6 +9,7 @@ from sqlalchemy.orm import joinedload
 
 from nyaa import backend, forms, models, torrents
 from nyaa.extensions import db
+from nyaa.template_utils import get_category_id_map
 from nyaa.utils import cached_function
 
 app = flask.current_app
@@ -251,7 +252,7 @@ def upload():
 def _create_upload_category_choices():
     ''' Turns categories in the database into a list of (id, name)s '''
     choices = [('', '[Select a category]')]
-    id_map = backend.get_category_id_map()
+    id_map = get_category_id_map()
 
     for key in sorted(id_map.keys()):
         cat_names = id_map[key]
